Skip showing artist seprately if browser song title contains artist name
This fails for some self titled songs played through the browser,
as well as any song that happens to contain the artists name in it.
It also fails to identify the artist if the spelling differs, such as
using different kanji or using kana in one and kanji in the other.
Though the code does normalise filtering all of those out didn't seem
plausible as there doesn't seem to be a simple heuristic to do so.